chore: add fram specific e2e test
